- removed the old tagger test - renamed SqlTest to TaggerTest
